test(server-mongo-testing): move tests for locally started MongoDB in dedicated test class
